chore: [#248] Move Subheader and Bloodline into subcomponent files
