+++ src/components/CowCard/Bloodline/Bloodline.js
@@ -1,0 +1,32 @@
+import React, { memo } from 'react'
+import { object } from 'prop-types'
+import Tooltip from '@material-ui/core/Tooltip'
+
+import { COW_COLOR_NAMES } from '../../../strings'
+
+import './Bloodline.sass'
+
+const Bloodline = ({ colorsInBloodline }) => (
+  <ul {...{ className: 'Bloodline' }}>
+    {Object.keys(colorsInBloodline)
+      .sort()
+      .map(color => (
+        <Tooltip
+          {...{
+            key: color,
+            arrow: true,
+            placement: 'top',
+            title: COW_COLOR_NAMES[color],
+          }}
+        >
+          <li {...{ className: color.toLowerCase() }} />
+        </Tooltip>
+      ))}
+  </ul>
+)
+
+Bloodline.propTypes = {
+  colorsInBloodline: object.isRequired,
+}
+
+export default memo(Bloodline)
