+++ src/components/CowCard/Bloodline/Bloodline.sass
@@ -1,0 +1,36 @@
+.CowCard
+  .Bloodline,
+  // This additional specificity variant is needed to override the global .card-list styles
+  .card-list & ul.Bloodline
+    display: flex
+    margin: 1em 0
+
+    li
+      $indicator-size: 1.25em
+      background: #000
+      border-radius: 50%
+      border: solid 1px #000
+      height: $indicator-size
+      margin-bottom: 0
+      margin-right: 0.5em
+      margin-top: 0
+      width: $indicator-size
+
+      &.blue
+        background: #8ff0f9
+      &.brown
+        background: #b45f28
+      &.green
+        background: #65f295
+      &.orange
+        background: #ff7031
+      &.purple
+        background: #d884f2
+      &.white
+        background: #fff
+      &.yellow
+        background: #fff931
+
+  &.is-purchased
+    ul.Bloodline
+      margin: 0 0 1em
