+++ src/components/CowCard/Bloodline/Bloodline.test.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+
+import Bloodline from './Bloodline'
+
+let component
+
+beforeEach(() => {
+  component = shallow(<Bloodline {...{ colorsInBloodline: {} }} />)
+})
+
+test('renders', () => {
+  expect(component).toHaveLength(1)
+})
