+++ src/components/CowCard/Bloodline/index.js
@@ -1,0 +1,1 @@
+export { default } from './Bloodline'
