+++ src/components/CowCard/Subheader/Subheader.sass
@@ -1,0 +1,7 @@
+@import ../../../styles/variables.sass
+
+.CowCard
+  .Subheader
+    .hearts
+      [data-prefix="fas"] // Full heart
+        color: $heart-color
